Feature: Add support for build profile (#555)
* Add support for build profile

The build profile is passed to cpt.packager.run() using the
base_profile_build_name parameter.

When running in Docker the complete build profile is passed to the
container using an environment variable and will there be stored in a
temporary file.

Unfortunately the current implementation does not allow passing
additional options, settings or environment variables to the build
profile like it is done for 'normal' profiles. This feature is not
supported by cpt.packager.ConanMultipackager.add().